% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdowneditor.R
\name{mdeditorInput}
\alias{mdeditorInput}
\title{Markdown Editor Input}
\usage{
mdeditorInput(id, label, markdown = "", config = list())
}
\arguments{
\item{id}{The input id}

\item{label}{The label for the input}

\item{markdown}{Initial markdown content (default: "")}

\item{config}{Configuration options for the markdown editor}
}
\description{
Creates a markdown editor input for Shiny applications using reactR
}
